+++ CHANGELOG.md
@@ -17,6 +17,7 @@
  - Disabled futility pruning when enemy king is checked
  - Added detection of open file next to the king in the evaluation
  - Fixed memory usage issues during search
+ - Added history heuristic veto in LMR
 
 # Version 4.0 (Komekko), 22.01.2021
  - Added Texel project